+++ python/shipStrawTracking.py
@@ -77,7 +77,7 @@ def run_track_pattern_recognition(input_file, geo_file, output_file, method):
     # Create dummy output file as the  input file is updated directly and
     # histograms are written to output file (hists.root by default)
     run.SetSink(ROOT.FairRootFileSink(ROOT.TMemFile('output', 'recreate')))
-    run.SetUserConfig("g4Config_basic.C") # geant4 transport not used, only needed for the mag field
+    ROOT.gInterpreter.ProcessLine('FairRunSim::Instance()->SetSimulationConfig(std::make_unique<SHiP::VMCConfig>("g4Config", "g4Config_basic.yaml"));') # geant4 transport not used, only needed for the mag field
     rtdb = run.GetRuntimeDb()
 
     modules = shipDet_conf.configure(run,ShipGeo)